[IMP] Moved last context route to namespace app.blueprints.rest
